+++ modules/Students/includes/StudentFields.php
@@ -6,10 +6,12 @@
  * @subpackage modules
  */
 
-if ( isset( $_POST['tables'] )
-	&& is_array( $_POST['tables'] )
+// @since 13.0 Add `&modfunc=save` to Fields form
+if ( $_REQUEST['modfunc'] === 'save'
 	&& AllowEdit() )
 {
+	$_REQUEST['tables'] = issetVal( $_REQUEST['tables'], [] );
+
 	$table = issetVal( $_REQUEST['table'] );
 
 	if ( ! in_array( $table, [ 'student_field_categories', 'custom_fields' ] ) )
@@ -138,8 +140,8 @@ if ( isset( $_POST['tables'] )
 			$error[] = _( 'Please enter valid Numeric data.' );
 	}
 
-	// Unset tables & redirect URL.
-	RedirectURL( 'tables' );
+	// Unset modfunc, tables & redirect URL.
+	RedirectURL( [ 'modfunc', 'tables' ] );
 }
 
 // Delete Field / Category.